+++ src/components/MyProfile.js
@@ -38,8 +38,10 @@ class MyProfile extends Component {
         })
       })
       .catch(res=>{
-        localStorage.clear();
-        window.location.href = "/login";
+        if (res.status === 401) {
+          localStorage.clear();
+          window.location.href = "/login";
+        }
       })
     }
     
@@ -84,8 +86,10 @@ class MyProfile extends Component {
        window.location.reload();
       })
       .catch(res => {
-        localStorage.clear();
+        if (res.status === 401) {
+          localStorage.clear();
           window.location.href = "/login";
+        }
       });
   };
