+++ src/services/api.js
@@ -15,8 +15,11 @@ export function apiCall(method, path, data) {
         return resolve(res.data);
       })
       .catch(err => {
-          localStorage.clear();
+          if(err.status === 401)
+          {
+            localStorage.clear();
             window.location.href = "/login";
+          }
       });
   });
 }
